requirements.txt
amqp==2.5.0
anyjson==0.3.3
attrs==19.1.0
Babel==2.7.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
bleach==3.1.0
blessings==1.7
boto3==1.9.183
botocore==1.12.183
bpython==0.18
bs4==0.0.1
cachetools==3.1.1
celery==4.3.0
celery-flower==1.0.1
certifi==2019.3.9
chardet==3.0.4
cloudinary==1.16.0
curtsies==0.3.0
decorator==4.4.0
demjson==2.2.4
Django==2.2.2
django-celery==3.3.0
django-celery-results==1.1.2
django-extensions==2.1.9
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
docutils==0.14
entrypoints==0.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
greenlet==0.4.15
httplib2==0.13.1
idna==2.8
ipython-genutils==0.2.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter-core==4.5.0
kombu==4.6.1
lxml==4.3.4
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.3.1
nbformat==4.4.0
oauthlib==3.1.0
pandocfilters==1.4.2
pika==1.0.1
Pillow==6.1.0
psycopg2==2.8.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
Pygments==2.4.2
PyJWT==1.7.1
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
sentry-sdk==0.10.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
testpath==0.4.2
titlecase==0.12.0
tornado==4.2
traitlets==4.3.2
typing==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
yadisk==1.2.14
